Api/lambdaLogic/inference/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
cycler==0.11.0
flatbuffers==23.3.3
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
monai==0.9.0
multidict==6.0.4
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-ignite==0.4.9
pytorch-lightning==1.6.4
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tifffile==2021.11.2
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.2
torchtext==0.12.0
torchvision==0.12.0
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
